auth: (bind) fix one bad case of time_t truncation

+++ modules/bindbackend/bindbackend2.cc
@@ -492,8 +492,7 @@ void Bind2Backend::getUnfreshSecondaryInfos(vector<DomainInfo>* unfreshDomains)
     catch (...) {
     }
     sd.serial = soadata.serial;
-    // coverity[store_truncates_time_t]
-    if (sd.last_check + soadata.refresh < (unsigned int)time(nullptr))
+    if (sd.last_check + soadata.refresh < time(nullptr))
       unfreshDomains->push_back(std::move(sd));
   }
 }
